% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomeUtils.R
\docType{package}
\name{biomeUtils-package}
\alias{biomeUtils-package}
\title{\code{biomeUtils} Package.}
\description{
\code{biomeUtils} is part of the RIVM Toolbox for Microbiome data analyses
\code{biomeUtils} implements utility tools for microbiota data formatting,
handling, and exploration.
}
