% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSeqTree.R
\name{buildSeqTree}
\alias{buildSeqTree}
\title{Build a phylogenetic tree}
\usage{
buildSeqTree(
  x,
  pml_model = "GTR",
  pml_optInv = TRUE,
  pml_optGamma = TRUE,
  pml_rearrangement = "stochastic",
  pml_control = pml.control(trace = 0)
)
}
\arguments{
\item{x}{Sequences DNAStringSet}

\item{pml_model}{See optim.pml \code{phangorn}}

\item{pml_optInv}{See optim.pml \code{phangorn}}

\item{pml_optGamma}{See optim.pml \code{phangorn}}

\item{pml_rearrangement}{See optim.pml \code{phangorn}
#'}

\item{pml_control}{See optim.pml \code{phangorn}}
}
\description{
Build a phylogenetic tree
}
\details{
A wrapper to build phylogenetic tree for ASV sequences
         obtained after \code{dada2}.
}
\examples{
library(biomeUtils)
data("SprockettTHData")
# select few for example
seqs <- sample(SprockettTHData@refseq, 5)
sp.tree <- buildSeqTree(seqs)
sp.tree$tree

}
