% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePD.R
\name{calculatePD}
\alias{calculatePD}
\title{Calculate Faith's Phylogenetic Diversity}
\usage{
calculatePD(x, justDF = FALSE, include_root = TRUE)
}
\arguments{
\item{x}{A phyloseq object}

\item{justDF}{Logical if TURE returns only sample data as tibble with
PD and SR values as columns.Default is FALSE}

\item{include_root}{Logical if root node be included. Depends whether the
input phyloseq had rooted tree (default = TRUE).}
}
\value{
Either a phyloseq object or data.frame with results of PD
        added to sample_data()
}
\description{
Calculate Faith's Phylogenetic Diversity
}
\details{
A wrapper around \code{picante} for \code{phyloseq} objects
         to calculate and add Phylogenetic Diversity and Species Richness
         to sample data.
}
\examples{
library(biomeUtils)
data("FuentesIliGutData")
# reduce size for example
ps1 <- subset_samples(FuentesIliGutData, ILI == "C")
ps1 <- prune_taxa(taxa_sums(ps1) > 0, ps1)

meta_tib <- calculatePD(ps1, justDF = TRUE)
# check
meta_tib[c(1, 2, 3), c("PD", "SR")]
}
\references{
\itemize{
\item{}{Kembel, S.W., Cowan, P.D., et al., 2010. Picante: R tools
       for integrating phylogenies and ecology.
       \emph{Bioinformatics}, 26(11), pp.1463-1464.
       \url{https://doi.org/10.1093/bioinformatics/btq166}
}
}
}
\author{
Sudarshan A. Shetty
}
