% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateQC.R
\name{calculateQC}
\alias{calculateQC}
\title{Calculate QC Metrics for Taxa and Samples}
\usage{
calculateQC(x)
}
\arguments{
\item{x}{A phyloseq object}
}
\value{
A list with two tibbles.

\emph{SampleQC}
\itemize{
\item{sample_total_taxa:}{Total taxa in a sample}
\item{sample_total_reads:}{Total reads in a sample}
\item{number_taxa_account_fifty_percent:}{Number of taxa account for 50 percent}
}
\emph{TaxaQC}
\itemize{
\item{taxa_counts:}{Total counts of taxa in all samples}
\item{taxa_mean_counts:}{Mean counts of taxa in all samples}
\item{taxa_dectected_samples:}{Number of samples in which taxa detected}
\item{taxa_prevalence:}{Percent of samples in which taxa detected}
}
}
\description{
Calculate QC Metrics for Taxa and Samples
}
\details{
Calculate QC metrics for taxa and samples.
}
\examples{
library(biomeUtils)
data("FuentesIliGutData")
qc_Data <- calculateQC(FuentesIliGutData)
qc_Data

}
\author{
Sudarshan A. Shetty
}
