% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateTaxaFoldDifference.R
\name{calculateTaxaFoldDifference}
\alias{calculateTaxaFoldDifference}
\title{Calculate Fold Difference in Taxa Abundance}
\usage{
calculateTaxaFoldDifference(x, group, sort = FALSE, paired = FALSE)
}
\arguments{
\item{x}{A phyloseq object}

\item{group}{Vector with specifying the groups to compare.
Only two-group comparisons are supported.}

\item{sort}{Sort the results by descending order of fold difference}

\item{paired}{Paired comparison (Default: FALSE)}
}
\value{
A tibble with taxa ids, taxonomic information,
        two-group prevalence and fold change values.
}
\description{
Calculate Fold Difference in Taxa Abundance
}
\details{
Calculate Log10 fold difference in abundance of taxa between two-
         groups. This code is modified from original code
         \url{https://github.com/microbiome/microbiome/blob/fa2c0de2fbe000da87be3c185972ed7f0f626591/inst/extdata/check_foldchange.R} Get the prevalence of taxa in \code{phyloseq} objects
         along with taxonomic classification and prevalence.
}
\examples{
library(biomeUtils)
data("FuentesIliGutData")
# reduce size for example
ps1 <- filterSampleData(FuentesIliGutData, ILI != "L2")

taxa_fd <- calculateTaxaFoldDifference(ps1, group = "ILI")
# check
taxa_fd
}
\references{
Shetty SA (2021). Utilities for microbiome analytics.
\url{https://github.com/RIVM-IIV-Microbiome/biomeUtils}
}
\author{
Original Author: Leo Lahti. Adapted by Sudarshan A. Shetty
}
