% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPolyphyletic.R
\name{checkPolyphyletic}
\alias{checkPolyphyletic}
\title{Check for Polyphyletic Taxa}
\usage{
checkPolyphyletic(x, taxa_level = "Genus", return_df = TRUE)
}
\arguments{
\item{x}{\code{phyloseq} object}

\item{taxa_level}{Taxonomic level to check. Default is Genus}

\item{return_df}{Logical. Return data.frame or list}
}
\value{
Data frame or list
}
\description{
Check for Polyphyletic Taxa
}
\details{
Check for polyphyletic taxa in \code{tax_table}. Useful to check this
         before aggregating at any level.
}
\examples{
library(biomeUtils)
library(dplyr)
data("FuentesIliGutData")
polydf <- checkPolyphyletic(FuentesIliGutData)
polydf

}
\references{
Shetty SA (2020). Utilities for microbiome analytics.
\url{https://github.com/RIVM-IIV-Microbiome/biomeUtils}
}
\author{
Sudarshan A. Shetty
}
