% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFilterEffect.R
\name{chekFilterEffect}
\alias{chekFilterEffect}
\alias{checkAbundanceCutOffs}
\alias{checkPrevalenceCutOffs}
\title{Check Filter Effect}
\usage{
checkAbundanceCutOffs(x, abundance_thres = NULL, prevalence_thres = NULL)

checkPrevalenceCutOffs(x, abundance_thres = NULL, prevalence_thres = NULL)
}
\arguments{
\item{x}{\code{phyloseq} object}

\item{abundance_thres}{If using 'checkAbundanceCutOffs', can have multiple
numeric values like c(0.1, 0.01, 0.001, 0.0001).
If using 'checkPrevalenceCutOffs', a single value.}

\item{prevalence_thres}{If using 'checkPrevalenceCutOffs', can have multiple
numeric values like c(0.05, 0.1, 0.15, 0.2).
If using 'checkAbundanceCutOffs', a single value.}
}
\value{
A tibble
}
\description{
Check Filter Effect
}
\details{
Investigate effect of abundance and prevalence thresholds on
         taxa numbers.

\code{checkAbundanceCutOffs} check how many taxa are lost at different
abundance thresholds.

\code{checkPrevalenceCutOffs} check how many taxa are lost at different
prevalence thresholds.
}
\examples{
library(biomeUtils)
library(microbiome)
data("FuentesIliGutData")
ps.rel <- microbiome::transform(FuentesIliGutData, "compositional")
ab.check <- checkAbundanceCutOffs(ps.rel,
                                  abundance_thres = c(0.0001, 0.001, 0.01, 0.1),
                                  prevalence_thres = 0.1)
ab.check

pv.check <- checkPrevalenceCutOffs(ps.rel,
                                   abundance_thres = 0.01,
                                   prevalence_thres = c(0.05, 0.1, 0.15, 0.2))
pv.check

}
\references{
Shetty SA (2020). Utilities for microbiome analytics.
\url{https://github.com/microsud/biomeUtils}
}
\author{
Sudarshan A. Shetty
}
