% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanTaxonomy.R
\name{cleanTaxonomy}
\alias{cleanTaxonomy}
\title{Clean Taxonomy}
\usage{
cleanTaxonomy(x)
}
\arguments{
\item{x}{A phyloseq object}
}
\value{
phyloseq
}
\description{
Clean Taxonomy
}
\details{
Removes, special characters, spaces and prefixes in tax_table.
}
\examples{
library(biomeUtils)
data("FuentesIliGutData")
ps1 <- cleanTaxonomy(FuentesIliGutData)
ps1
}
\author{
Sudarshan A. Shetty
}
