% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterASVs.R
\name{clusterASVs}
\alias{clusterASVs}
\title{Cluster ASVs into Clusters}
\arguments{
\item{x}{A phyloseq object}

\item{cluster_method}{See methods in \code{\link[DECIPHER]{IdClusters}}
in \code{DECIPHER}
Default is "inexact"}

\item{cluster_cutoff}{Cut-off for clustering ASVs.
Default= 0.01 i.e. 99 percent identity}

\item{threads}{No. of processors to use}
}
\value{
A new \code{phyloseq} object with cluster ASVs
}
\description{
Cluster ASVs into Clusters
}
\details{
clusterASVs clusters ASV sequences
         obtained after \code{dada2} in \code{phyloseq} object into clusters
         based on user provided sequence identity. If using method="inexact"
         this is similar to OTU approach for example clustering of ASVs that
         are 97% similar into one operational taxonomic unit. While preference
         for ASVs or OTUs is highly debated (see reference section), both
         have their pros and cons.The user is advised to make appropriate
         choice based on their research questions and properties of data.
}
\examples{
\dontrun{
library(biomeUtils)
library(dplyr)
library(data.table)
library(tibble)

testSprockett <- core(subset_samples(SprockettTHData, Community == "River_4"), 5, 20 / 100)
testSprockett <- clusterASVs(testSprockett,
                             cluster_cutoff = 0.01,
                             threads = 2)
}
}
\references{
\itemize{
\item{}{Clustering ASVs after DADA2? generating new OTU table.
       \emph{GitHub issue #947:}
       \url{https://github.com/benjjneb/dada2/issues/947}
}
\item{}{Michael McLaren. mikemc/speedyseq: speedyseq v0.2.0.
        \emph{Zenodo}, 2020, June 30, (Version v0.2.0).
        \url{http://doi.org/10.5281/zenodo.3923184}
}
\item{}{PD Schloss. Pre-print Review of 97% Identity Threshold Manuscript.
        \emph{Blogpost: The Academic Hermit}, 2017, Oct 11.
        \url{http://www.academichermit.com/2017/10/11/Review-of-Updating-the-97-identity-threshold.html}
}
\item{}{Angela Oliverio and Noah Fierer. Intragenomic heterogeneity and its
        implications for ESVs.
        \emph{Blogpost: Fierer Lab}, 2017, Oct 9.
        \url{http://fiererlab.org/2017/10/09/intragenomic-heterogeneity-and-its-implications-for-esvs/}
}
\item{}{Noah Fierer, Tess Brewer, and Mallory Choudoir. Lumping versus
        splitting – is it time for microbial ecologists to abandon OTUs?.
        \emph{Blogpost: Fierer Lab}, 2017, May 2.
        \url{http://fiererlab.org/2017/05/02/lumping-versus-splitting-is-it-time-for-microbial-ecologists-to-abandon-otus/}
}
}
}
\author{
Sudarshan A. Shetty
}
