% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dysbiosisNBC.R
\name{dysbiosisScore}
\alias{dysbiosisScore}
\title{Neighborhood Classification Dysbiosis Scores}
\usage{
dysbiosisScore(x, reference_samples = NULL, method = c("median-BC", "BC-ED"))
}
\arguments{
\item{x}{A phyloseq object}

\item{reference_samples}{Vector of samples to use as reference.}

\item{method}{Either 'median-BC' or 'BC-ED'}
}
\value{
A tibble data frame with dysbiosis score and sample information
}
\description{
Neighborhood Classification Dysbiosis Scores
}
\details{
Calculates neighborhood classification based microbiome dysbiosis scores.

Dysbiosis score or index are often used to identify the extent to which
clinically classified diseased samples are different from non diseased
reference samples. There are several approaches large-scale bacterial marker
profiling, relevant taxon-based methods, neighborhood classification,
random forest prediction, and combined alpha-beta diversity reviewed by Wei 2021.
Here we provide two scores based on neighborhood classification.

\itemize{

\item{'median-BC' }{The median Bray-Curtis distance between the test
sample and the reference samples (Lloyd-Price J, Arze C, Ananthakrishnan AN
et al. 2021).}

\item{'BC-ED' }{The Euclidian distance between test sample and the centroid (median)
of PCoA Axis 1 and 2 (based on Bray-Curtis dissimilarity) of the reference
samples. (Mahapatra S et al 2021)}

}
}
\examples{
library(biomeUtils)
data("FuentesIliGutData")
# Keep only two groups. controls and L1.
ps <- filterSampleData(FuentesIliGutData, ILI != "L2")
ps <- getProportions(ps)
# Define controls as reference samples
ref_samples <- rownames(meta(subset_samples(ps, ILI == "C")))

dys_bc <- dysbiosisScore(ps,
                         reference_samples = ref_samples,
                         method="median-BC")
dys_bc_ed <- dysbiosisScore(ps,
                            reference_samples = ref_samples,
                            method="BC-ED")

}
\references{
\itemize{
\item{}{Lloyd-Price J, Arze C, Ananthakrishnan AN et al. (2019).
Multi-omics of the gut microbial ecosystem in inflammatory bowel diseases.
\emph{Nature}, 569(7758), pp.655-662.}

\item{}{Mahapatra S et al. (2021). Nanopore 16S rRNA sequencing reveals
alterations in nasopharyngeal microbiome and enrichment of Mycobacterium
and Mycoplasma in patients with COVID 19.
\emph{medRxiv preprint}, 2021.11.10.21266147.}

}
}
\author{
Sudarshan A. Shetty
}
