% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchSequence.R
\name{fetchSequence}
\alias{fetchSequence}
\title{Retrieve Sequence from Phyloseq}
\usage{
fetchSequence(x, seq_id = NULL, as_DNAStringSet = TRUE)
}
\arguments{
\item{x}{A phyloseq object}

\item{seq_id}{A list of IDs matching taxa_names(phyloseq) that also have
sequence information stored in refseq slot of
\code{phyloseq} object.}

\item{as_DNAStringSet}{Logical (Default=TRUE). If FALSE sequences are
returned as character strings.}
}
\value{
Sequences for user species IDs either as vector
        or DNAStringSet object
}
\description{
Retrieve Sequence from Phyloseq
}
\details{
Retrieve sequences from refseq slot in phyloseq
         \code{phyloseq} object.
}
\examples{
library(biomeUtils)
data("SprockettTHData")
ps <- SprockettTHData
seq_id <- c("ASV_30_Lachnospiraceae", "ASV_451_UC5-1-2E3", "ASV_586_CAG-56")
myseqs <- fetchSequence(ps, seq_id = seq_id, as_DNAStringSet = TRUE)
myseqs
}
\author{
Sudarshan A. Shetty
}
