% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterPhyloseq.R
\name{filterPhyloseq}
\alias{filterPhyloseq}
\alias{filterSampleData}
\alias{filterTaxaData}
\alias{filterTaxaByNames}
\alias{filterSampleByNames}
\title{Filter Phyloseq Object}
\usage{
filterSampleData(x, ...)

filterTaxaData(x, ...)

filterTaxaByNames(x, ids = NULL, keep = TRUE)

filterSampleByNames(x, ids = NULL, keep = TRUE)
}
\arguments{
\item{x}{\code{phyloseq} object}

\item{...}{Option to pass on to \code{\link[dplyr]{filter}}. Only when using
\code{filterSampleData} or \code{filterTaxaData}
sample data or taxonomy table as tibble.}

\item{ids}{A list of sample ids or taxa ids to filter. Only when using
\code{filterSampleByNames} or \code{filterTaxaByNames}}

\item{keep}{Logical. Default is TRUE. Only when using
\code{filterSampleByNames} or \code{filterTaxaByNames}}
}
\value{
Either filtered \code{phyloseq}
}
\description{
Filter Phyloseq Object
}
\details{
These are alternative to subset and prune functions in
         \code{phyloseq}.

The \code{filterSampleData} does subsetting similar to
\code{\link[phyloseq]{subset_samples}}.
The \code{filterTaxaData} does subsetting similar to
\code{\link[phyloseq]{subset_taxa}}.


The \code{filterSampleByNames} does pruning similar to
\code{\link[phyloseq]{prune_samples}}.
The \code{filterTaxaByNames} does pruning similar to
\code{\link[phyloseq]{prune_taxa}}.

The naming is using filter because it uses the \code{dplyr::filter} function.

Note:
\code{filterSampleData} will additionally remove any taxa that are
zero across all samples i.e. not present is any remaining samples
after filtering.
\code{filterTaxaData} will additionally remove samples that do not
have any of the selected taxa, i.e. samples that sum to zero.
\code{filterSampleByNames} will additionally remove any taxa that are not
present is any remaining samples
after filtering.
\code{filterTaxaByNames} will additionally remove samples that do not
have any of the selected taxa, i.e. samples that sum to zero.
}
\examples{

library(biomeUtils)
data("FuentesIliGutData")

# Filter from tables subset_*
ps.filtered.samples  <- filterSampleData(FuentesIliGutData,
                                         ILI == "C" & BMI < 26)

ps.filtered.samples

ps.filtered.taxa  <- filterTaxaData(FuentesIliGutData,
                                    Phylum=="Firmicutes")

ps.filtered.taxa

# Filter by names like prune_*
sams.select <- sample_names(FuentesIliGutData)[1:10]
ps.filter.by.sam.names <- filterSampleByNames(FuentesIliGutData,
                                              ids = sams.select,
                                              keep = TRUE)

tax.select <- taxa_names(FuentesIliGutData)[1:10]
ps.filter.by.tax.names <- filterTaxaByNames(FuentesIliGutData,
                                            ids = tax.select,
                                            keep = TRUE)

}
\references{
Shetty SA (2020). Utilities for microbiome analytics.
\url{https://github.com/microsud/biomeUtils}
}
\author{
Sudarshan A. Shetty
}
