% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterTaxaTotal.R
\name{filterTaxaTotal}
\alias{filterTaxaTotal}
\title{Filter Taxa Based on Total Data}
\arguments{
\item{x}{\code{phyloseq} object.}

\item{rank}{Taxonomic rank to use like Phylum.}

\item{percent_thres}{Percent cut-off to use. If value is 10, then all phyla with
less that 10 percent of the total counts in all data are removed.}

\item{verbose}{Logical. Prints a list of removed taxa. Default is TRUE.}
}
\value{
Filtered phyloseq
}
\description{
Filter Taxa Based on Total Data
}
\details{
Provided a count data \code{phyloseq}, \code{filterTaxaTotal} calculates
         the percent of taxon at the rank level is calculated. Those that are
         less than the percent_thres are removed. This function works with ASV
         level data as the ASVs are merged at specified level for calculation.
         Therefore, ASVs that belong to low/rare abundance at a specified rank
         are removed.
}
\examples{
library(biomeUtils)
data('FuentesIliGutData')
# below we filter Family that are less than 2\% of the total data
ps.filt <- filterTaxaTotal(FuentesIliGutData,
                           rank = 'Family',
                           percent_thres = 2,
                           verbose = TRUE)
ps.filt

}
\author{
Sudarshan A. Shetty
}
