% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTopTaxa.R
\name{findTopTaxa}
\alias{findTopTaxa}
\title{Find the Top Taxa}
\usage{
findTopTaxa(x, top = 5L, method = c("mean", "sum", "median", "prevalence"))
}
\arguments{
\item{x}{A phyloseq object}

\item{top}{Numeric value, how many top taxa to return. Default return top
five taxa.}

\item{method}{Specify the method to determine top taxa. Either sum, mean,
median or prevalence. Default is 'mean'.}
}
\value{
A vector of the most \code{top} abundant taxa
}
\description{
Find the Top Taxa
}
\examples{
library(biomeUtils)
data("FuentesIliGutData")
findTopTaxa(FuentesIliGutData, top= 5L, method="prevalence")

}
\references{
Shetty SA (2020). Utilities for microbiome analytics.
\url{https://github.com/microsud/biomeUtils}
}
\author{
Sudarshan A. Shetty
}
