% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FuentesIliGutData}
\alias{FuentesIliGutData}
\title{FuentesIliGutData}
\format{
Phyloseq object
}
\usage{
data(FuentesIliGutData)
}
\value{
Loads the data set in R.
}
\description{
Fuentes S, et al. (2021) data in which the authors investigated
             associations between gut microbiota in respiratory infections and
             local and systemic immune markers in older adults during an influenza
             season.
}
\details{
FuentesIliGutData consists of 905 taxa (ASV-level) in 589 samples from
         397 participants. The samples include follow-up (recovery) samples
         that were collected 7–9 weeks after the ILI event.
         Data on co-morbidities, medications, symptoms, pathogens are made
         available.
}
\references{
\itemize{
\item{}{Fuentes S, et al. (2021) Associations of faecal microbiota
       with influenza-like illness in participants aged 60 years or
       older: an observational study.
       \emph{The Lancet Healthy Longevity}, 2(1):e13-23.
       \url{https://doi.org/10.1016/S2666-7568(20)30034-9}
}
\item{}{Ferreira JA, Fuentes S. (2020). Some comments on certain
        statistical aspects of the study of the microbiome.
        \emph{Briefings in bioinformatics} 21(4), pp.1487-1494.
        \url{https://doi.org/10.1093/bib/bbz077}
}
}
}
\author{
Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Data}
