% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericUtilities.R
\name{genericUtilities}
\alias{genericUtilities}
\alias{meltAbundance}
\alias{transposeAbundance}
\title{Generic utilities}
\usage{
meltAbundance(x, ...)

transposeAbundance(x, transform = "identity")
}
\arguments{
\item{x}{\code{phyloseq} object}

\item{...}{Arguments other than 'cols' to pass to tidyr pivot_longer function}

\item{transform}{If transformation is required `transposeAbundance`}
}
\value{
long data format.
}
\description{
Generic utilities
}
\details{
Generic utilities that are simple wrappers for doing common
         tasks.

\code{meltAbundance} convert abundance data to long format.

\code{transposeAbundance} transpose OTU for use with \code{vegan} pkg.
}
\examples{

library(biomeUtils)
data("FuentesIliGutData")
melt.abund <- meltAbundance(FuentesIliGutData,
                            names_to = "SampleID",
                            values_to="Abundance")
melt.abund

t.otu <- transposeAbundance(FuentesIliGutData)
dim(t.otu)
}
\references{
Shetty SA (2020). Utilities for microbiome analytics.
\url{https://github.com/RIVM-IIV-Microbiome/biomeUtils}
}
\author{
Sudarshan A. Shetty
}
