% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhyloseqSlots.R
\name{getPhyloseqSlots}
\alias{getPhyloseqSlots}
\alias{getOtuTableFromTibble}
\alias{getTaxaTableFromTibble}
\alias{getSampleTableFromTibble}
\title{Get Phyloseq slots from Tibbles}
\usage{
getOtuTableFromTibble(x, rownames = "FeatureID")

getTaxaTableFromTibble(x, rownames = "FeatureID")

getSampleTableFromTibble(x, rownames = NULL)
}
\arguments{
\item{x}{\code{phyloseq} object}

\item{rownames}{A column in tibble to use are rownames}
}
\value{
Either \code{\link[phyloseq]{otu_table}} or
               \code{\link[phyloseq]{tax_table}} or
               \code{\link[phyloseq]{sample_data}}
}
\description{
Get Phyloseq slots from Tibbles
}
\details{
Convert different tibbles into \code{phyloseq} slots.

\code{getOtuTableFromTibble} Converts OTU tibble obtained using
\code{\link[biomeUtils]{getAbundanceTibble}} to \code{\link[phyloseq]{otu_table}}.

\code{getTaxaTableFromTibble} Converts the taxa tibble obtained using
\code{\link[biomeUtils]{getTaxaTibble}} to \code{\link[phyloseq]{tax_table}}.

\code{getSampleTableFromTibble} Converts the sample data tibble obtained using
\code{\link[biomeUtils]{getSampleTibble}} to \code{\link[phyloseq]{sample_data}}.
}
\examples{
library(biomeUtils)
data("FuentesIliGutData")
otu_tib  <- getAbundanceTibble(FuentesIliGutData)
otu_table <- getOtuTableFromTibble(otu_tib, rownames="FeatureID")
class(otu_table)

tax_tib <- getTaxaTibble(FuentesIliGutData)
taxa_table <- getTaxaTableFromTibble(tax_tib, rownames="FeatureID")
class(tax_table)

meta_tib <- getSampleTibble(FuentesIliGutData)
sample_table <- getSampleTableFromTibble(meta_tib, rownames="SampleID")
class(sample_table)

}
\references{
Shetty SA (2020). Utilities for microbiome analytics.
\url{https://github.com/microsud/biomeUtils}
}
\author{
Sudarshan A. Shetty
}
