% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrevalence.R
\name{getPrevalence}
\alias{getPrevalence}
\title{Get Prevalence and Taxonomy}
\usage{
getPrevalence(
  x,
  return_rank = rank_names(x),
  return_taxa = NULL,
  sort = TRUE,
  ...
)
}
\arguments{
\item{x}{A phyloseq object}

\item{return_rank}{Specify which taxonomic ranks to include in output.
Must be a character vector \code{phyloseq::rank_names()}}

\item{return_taxa}{A specific list of taxa for which the values should
be returned. This can be used if user is not interested
in all the taxa in input \code{phyloseq}. Default is NULL
which returns all taxa. This list must match rows names if
otu_table in phyloseq has taxa_are_rows=TRUE or columns
names if otu_table in phyloseq has taxa_are_rows=FALSE}

\item{sort}{Logical. Sort by prevalence value from higher to lower (Default=TRUE)}

\item{...}{Option to pass microbiome::prevalence}
}
\value{
A tibble with prevalence and taxonomy
}
\description{
Get Prevalence and Taxonomy
}
\details{
Get the prevalence of taxa in \code{phyloseq} objects
         along with taxonomic classification.
}
\examples{
library(biomeUtils)
data("FuentesIliGutData")
prev_tib <- getPrevalence(FuentesIliGutData,
  return_rank = c("Family", "Genus"),
  return_taxa = c("ASV4", "ASV17", "ASV85", "ASV83"),
  sort = TRUE
)
head(prev_tib)
}
\references{
\itemize{
\item{}{A Salonen et al. (2012) The adult intestinal core microbiota
        is determined by  analysis depth and health status.
       \emph{Clinical Microbiology and Infection}, 18(S4):16 20.
       \url{https://doi.org/10.1111/j.1469-0691.2012.03855.x}
}
\item{}{Lahti L, Shetty S (2012-2019). microbiome R package.
        statistical aspects of the study of the microbiome.
        \emph{BioConductor}.
        \url{https://doi.org/doi:10.18129/B9.bioc.microbiome}
}
}
}
\author{
Sudarshan A. Shetty
}
