% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTaxaOverview.R
\name{getTaxaOverview}
\alias{getTaxaOverview}
\title{Get Overview of Taxa Abundances}
\usage{
getTaxaOverview(x, rank = NULL)
}
\arguments{
\item{x}{A phyloseq object}

\item{rank}{Default is NULL and returns overview matching \code{taxa_names}.
Else user can provide one of the ranks available in 'rank_names'
of phyloseq object.}
}
\value{
A \code{tibble}
}
\description{
Get Overview of Taxa Abundances
}
\examples{
library(biomeUtils)
data("FuentesIliGutData")
getProportions(FuentesIliGutData) |>
   filterSampleData(ILI == "C") |>
   getTaxaOverview(rank="Phylum")

}
\references{
Shetty SA (2020). Utilities for microbiome analytics.
\url{https://github.com/RIVM-IIV-Microbiome/biomeUtils}
}
\seealso{
getTaxaSummary
}
\author{
Sudarshan A. Shetty
}
