% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTaxaSummary.R
\name{getTaxaSummary}
\alias{getTaxaSummary}
\title{Get Taxa Summary at Specific Taxonomic Level}
\usage{
getTaxaSummary(x, rank = "Phylum")
}
\arguments{
\item{x}{A phyloseq object}

\item{rank}{Taxonomic rank for which abundance summaries
are required.}
}
\description{
Get Taxa Summary at Specific Taxonomic Level
}
\details{
Get a summary of total abundances of taxa
         at specific taxonomic level.
}
\examples{
library(biomeUtils)
library(dplyr)
data("SprockettTHData")
x <- subset_samples(SprockettTHData,
                    Delivery_Mode=="Vaginal" & Cohort == "Tsimane")
x <- removeZeros(x)
getTaxaSummary(x, rank="Phylum")

}
