% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTibbles.R
\name{getTibble}
\alias{getTibble}
\alias{getAbundanceTibble}
\alias{getTaxaTibble}
\alias{getSampleTibble}
\title{Get tibble}
\usage{
getAbundanceTibble(
  x,
  column_id = "FeatureID",
  select_rows = NULL,
  select_cols = NULL
)

getTaxaTibble(
  x,
  column_id = "FeatureID",
  select_rows = NULL,
  select_cols = NULL
)

getSampleTibble(
  x,
  column_id = "SampleID",
  select_rows = NULL,
  select_cols = NULL
)
}
\arguments{
\item{x}{\code{phyloseq} object}

\item{column_id}{A character}

\item{select_rows}{Rows to return in output.}

\item{select_cols}{Columns to return in output}
}
\value{
A tibble
}
\description{
Get tibble
}
\details{
Convert different \code{phyloseq} slots into tibbles.

\code{getAbundanceTibble} gets the otu_table in tibble format.

\code{getTaxaTibble} gets the taxa_table in tibble format.

\code{getSampleTibble} gets the sample_data in tibble format.
}
\examples{
library(biomeUtils)
data("FuentesIliGutData")
otu_tib <- getAbundanceTibble(FuentesIliGutData)
tax_tib <- getTaxaTibble(FuentesIliGutData)
meta_tib <- getSampleTibble(FuentesIliGutData)
}
\references{
Shetty SA (2020). Utilities for microbiome analytics.
\url{https://github.com/RIVM-IIV-Microbiome/biomeUtils}
}
\author{
Sudarshan A. Shetty
}
