% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYCTheta.R
\name{getYCTheta}
\alias{getYCTheta}
\title{Yue and Clayton Measure of Dissimilarity}
\arguments{
\item{x}{A phyloseq object or a data.frame with taxa-columns and samples-rows}
}
\value{
Object of class dist
}
\description{
Yue and Clayton Measure of Dissimilarity
}
\details{
Calculates Yue and Clayton measure of dissimilarity between samples.
         It considers total number of taxa between two samples and the
         relative abundance of each taxa.
}
\examples{
library(biomeUtils)
data("FuentesIliGutData")
yc.dis <- getYCTheta(FuentesIliGutData)
#yc.dis[1:3,1:4]
}
\references{
\itemize{
\item{}{Yue, Jack C., and Murray K. Clayton. 2010. A similarity measure
        based on species proportions.
       \emph{Communications in Statistics-theory and Methods}, 34(11), pp.2123-2131.
       \url{https://doi.org/10.1080/STA-200066418}
}
{Dixon, P 2003. VEGAN, a package of R functions for community ecology.
       \emph{Journal of Vegetation Science}, 14(6), pp.927-930.
       \url{https://doi.org/10.1111/j.1654-1103.2003.tb02228.x}
}
}
}
\author{
Sudarshan A. Shetty
}
