% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltDistanceToTable.R
\name{meltDistanceToTable}
\alias{meltDistanceToTable}
\title{Convert Distance matrix to Data Frame}
\usage{
meltDistanceToTable(x, dist_mat = NULL, name_dist_column = "value", ...)
}
\arguments{
\item{x}{A phyloseq object}

\item{dist_mat}{An object of class \code{dist}}

\item{name_dist_column}{A specific name for column where distance values
are stored. Default="value"}

\item{...}{Option to pass biomeUtils::getSampleTibble like column name and columns
to be included in the output.}
}
\value{
A data frame with pairwise distance and sample information
}
\description{
Convert Distance matrix to Data Frame
}
\details{
Converts \code{dist} object from \code{phyloseq::distance()}
         to a data frame combined with sample data
}
\examples{
library(biomeUtils)
data("FuentesIliGutData")
ps <- filterSampleData(FuentesIliGutData, ILI != "L2")
ps <- phyloseq::rarefy_even_depth(ps)
dist.mat <- phyloseq::distance(ps, "bray")
dist.melt.sample <- meltDistanceToTable(ps,
                                        dist_mat = dist.mat,
                                        name_dist_column = "Bray-Curtis",
                                        select_cols = c("participant_id", "ILI"))
head(dist.melt.sample)

}
\author{
Sudarshan A. Shetty
}
