% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutatePhyloseq.R
\name{mutatePhyloseq}
\alias{mutatePhyloseq}
\alias{mutateSampleData}
\alias{mutateTaxaTable}
\alias{mutateAbundanceTable}
\title{Mutating Slots of Phyloseq}
\usage{
mutateSampleData(x, ...)

mutateTaxaTable(x, ...)

mutateAbundanceTable(x, ...)
}
\arguments{
\item{x}{\code{phyloseq} object}

\item{...}{Arguments that dplyr::mutate can accept}
}
\value{
A phyloseq
}
\description{
Mutating Slots of Phyloseq
}
\details{
Mutate different \code{phyloseq} slots.

\code{mutateAbundanceTable} mutate \code{otu_table}. Will also add the new sample to
 \code{sample_data}

\code{mutateTaxaTable} mutate \code{tax_table}, Add or modify existing columns.

\code{mutateSampleData} mutate \code{sample_data}. Add columns, modify existing
columns.
}
\examples{
library(biomeUtils)
library(dplyr)
data("FuentesIliGutData")
ps_otu <- mutateAbundanceTable(FuentesIliGutData, test = 1:ntaxa(FuentesIliGutData))
ps_tax <- mutateTaxaTable(FuentesIliGutData, test = 1:ntaxa(FuentesIliGutData))
ps_samdat <- mutateSampleData(FuentesIliGutData, test = paste0(ILI, "-test"))
}
\references{
Shetty SA (2020). Utilities for microbiome analytics.
\url{https://github.com/microsud/biomeUtils}
}
\author{
Sudarshan A. Shetty
}
