% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullPhyloseq.R
\name{pullPhyloseq}
\alias{pullPhyloseq}
\alias{pullSampleAbundance}
\alias{pullTaxaAbundance}
\alias{pullTaxaRank}
\title{Pull Information from Phyloseq Object}
\usage{
pullSampleAbundance(x, ...)

pullTaxaAbundance(x, ...)

pullTaxaRank(x, ...)
}
\arguments{
\item{x}{\code{phyloseq} object}

\item{...}{Option to pass on to \code{\link[dplyr]{pull}}. Only when using
\code{filterSampleData} or \code{filterTaxaData}
sample data or taxonomy table as tibble.}
}
\value{
A vector of requested value
}
\description{
Pull Information from Phyloseq Object
}
\details{
These are alternative to subset and prune functions in
         \code{phyloseq}.

The \code{pullSampleAbundance} can be used to get abundances of all taxa.

The \code{pullTaxaAbundance} can be used to get abundances of a taxa in all
samples.

The \code{pullSampleVariable} can be used to get all variables within a
column in \code{sample_data}.

The \code{pullTaxaRank} can be used to get all taxomic information at a
particular rank in \code{tax_table}.

The naming is using pull because it uses the \code{dplyr::pull} function.
}
\examples{

library(biomeUtils)
data("FuentesIliGutData")
pullSampleAbundance(FuentesIliGutData, "sample_1", name= "FeatureID")

pullTaxaAbundance(FuentesIliGutData, "ASV1", name = "SampleID")

pullTaxaRank(FuentesIliGutData, "Genus")

}
\references{
Shetty SA (2020). Utilities for microbiome analytics.
\url{https://github.com/RIVM-IIV-Microbiome/biomeUtils}
}
\author{
Sudarshan A. Shetty
}
