% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMergedMetaphlan.R
\name{readMergedMetaphlan}
\alias{readMergedMetaphlan}
\title{Read Merged Metaphlan Table}
\usage{
readMergedMetaphlan(
  input_file_path = NULL,
  find_sample_name_pattern = NULL,
  replace_sample_name_pattern = "",
  return_species = TRUE
)
}
\arguments{
\item{input_file_path}{path to where the output of Metaphlan3 is located.}

\item{find_sample_name_pattern}{Naming convention used for samples within
the input file.}

\item{replace_sample_name_pattern}{specific pattern to replace in sample names.}

\item{return_species}{Logical whether to return only species level data. Default TRUE.}
}
\value{
A \code{phyloseq} object
}
\description{
Read Merged Metaphlan Table
}
\details{
This function reads Metaphlan3 output that is merged with all samples.
         and returns a \code{phyloseq} object with otu_table and tax_table.

Note:
\code{readMergedMetaphlan} is tested for Metaphlan version 3 and based on the RIVM
HPC HUMANN3 pipeline output built by Jeron Frank.
}
\examples{
\dontrun{

library(biomeUtils)

mpa.ps <- readMergedMetaphlan(input_file_path = "../metaphlan_bugs_list_merged.txt",
                             find_sample_name_pattern = "_kneaddata_concat_metaphlan_bugs_list",
                             replace_sample_name_pattern = "",
                             return_species = TRUE)
}

}
\references{
Shetty SA (2020). Utilities for microbiome analytics.
\url{https://github.com/RIVM-IIV-Microbiome/biomeUtils}
}
\author{
Sudarshan A. Shetty
}
