% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeZeros.R
\name{removeZeros}
\alias{removeZeros}
\title{Remove Zeros}
\usage{
removeZeros(x)
}
\arguments{
\item{x}{A phyloseq object or a data.frame with only numerical or matrix}
}
\value{
If input is phyloseq, returns phyloseq with taxa that
        have zeros in all samples.
        If input is matrix or data frame, returned object is
        without rows and cloumns that sum to zero.
}
\description{
Remove Zeros
}
\details{
In phyloseq based analysis, prune_samples or subset_samples
         does not remove taxa that have zero abundances in remaining
         samples. In such cases user may want to exclude these taxa.
}
\examples{
library(biomeUtils)
data("FuentesIliGutData")
ps1 <- subset_samples(FuentesIliGutData, ILI != "L2")
ps1
ps2 <- removeZeros(ps1)
ps2
}
\author{
Sudarshan A. Shetty
}
