% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameSampleVariables.R
\name{renameSampleVariables}
\alias{renameSampleVariables}
\title{Rename Sample Variables in Phyloseq Sample Data}
\usage{
renameSampleVariables(x, with = FALSE, ...)
}
\arguments{
\item{x}{\code{phyloseq} object}

\item{with}{Default is FALSE. Set this to TRUE when using a function. See
\code{\link[dplyr:rename_with]{rename_with}}}

\item{...}{Option to pass on to \code{\link[dplyr:rename]{rename}} or
\code{\link[dplyr:rename_with]{rename_with}}. If using later,
set with=TRUE.}
}
\value{
Phyloseq with modified sample_data
}
\description{
Rename Sample Variables in Phyloseq Sample Data
}
\details{
In some cases user may want to rename one or several sample variables.
         With \code{phyloseq}, this needs to be done one by one. To ease this step,
         we make use of the \code{\link[dplyr:rename]{rename}} and
         \code{\link[dplyr:rename_with]{rename_with}}. The later allows to use
         functions such as those that work with string matching.
}
\examples{
library(biomeUtils)
library(dplyr)
data(FuentesIliGutData)
FuentesIliGutData <- renameSampleVariables(FuentesIliGutData,
                                           with=FALSE,
                                           condition = "ILI")

FuentesIliGutData <- renameSampleVariables(FuentesIliGutData,
                                           with=TRUE,
                                           .fn = toupper)

}
\references{
Shetty SA (2020). Utilities for microbiome analytics.
\url{https://github.com/RIVM-IIV-Microbiome/biomeUtils}
}
\author{
Sudarshan A. Shetty
}
