% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveData.R
\name{saveData}
\alias{saveData}
\alias{saveAbundanceData}
\alias{saveTaxonomyData}
\alias{saveSampleData}
\title{Save data}
\usage{
saveAbundanceData(
  x,
  path_file = NULL,
  file_type = c("delim", "csv", "csv2", "tsv", "excel_csv", "excel_csv2"),
  ...
)

saveTaxonomyData(
  x,
  path_file = NULL,
  file_type = c("delim", "csv", "csv2", "tsv", "excel_csv", "excel_csv2"),
  ...
)

saveSampleData(
  x,
  path_file = NULL,
  file_type = c("delim", "csv", "csv2", "tsv", "excel_csv", "excel_csv2"),
  ...
)
}
\arguments{
\item{x}{\code{phyloseq} object}

\item{path_file}{path to save the file along with name.
Default is NULL. E.g. "data/output/otu_table.csv"}

\item{file_type}{One of "delim","csv","csv2","tsv","excel_csv","excel_csv2"
supported by \code{readr} package}

\item{...}{Arguments to pass to readr functions}
}
\value{
Store file locally specified by user.
}
\description{
Save data
}
\details{
Basic utilities for saving data in \code{phyloseq}.

\code{saveAbundanceData} save the otu_table.

\code{saveTaxonomyData} save the taxa_table.

\code{saveSampleData} save the sample_data.
}
\examples{
\dontrun{
library(biomeUtils)
data("FuentesIliGutData")
saveAbundanceData(FuentesIliGutData, path_file = "otu.tsv", file_type = "tsv")
saveTaxonomyData(FuentesIliGutData, path_file = "taxa.tsv", file_type = "tsv")
saveSampleData(FuentesIliGutData, path_file = "sample.tsv", file_type = "tsv")
}
}
\references{
Shetty SA (2020). Utilities for microbiome analytics.
\url{https://github.com/RIVM-IIV-Microbiome/biomeUtils}
}
\author{
Sudarshan A. Shetty
}
