% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveSequence.R
\name{saveSequence}
\alias{saveSequence}
\title{Save Sequence}
\usage{
saveSequence(x, file = NULL)
}
\arguments{
\item{x}{Either a phyloseq object with refseq information or
a object of class DNAStringSet.}

\item{file}{File name with extension ".fasta" or ".fa", e.g.
'myseqs.fasta' or file path where you wish to save the
output fasta e.g. 'mypath/myseqs.fasta'.
As a fail safe, default is NULL which returns
an error.}
}
\value{
A fasta file saved in user specified location
}
\description{
Save Sequence
}
\details{
Saves sequences either from a DNAStringSet or refseq slot
         \code{phyloseq} object. If taxa_names is sequences, use
         'add_refseq' function from \code{microbiomeutilties} before
         using this function.
}
\examples{

# data("SprockettTHData")
# ps <- SprockettTHData
# seq_id <- c("ASV_30_Lachnospiraceae", "ASV_451_UC5-1-2E3", "ASV_586_CAG-56")

# saveSequence(ps,
#             file="inst/extras/sequences.fasta")
}
\author{
Sudarshan A. Shetty
}
