% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separateSampleColumn.R
\name{separateSampleColumn}
\alias{separateSampleColumn}
\title{Separate a Column in Phyloseq Sample Data}
\usage{
separateSampleColumn(x, ...)
}
\arguments{
\item{x}{\code{phyloseq} object}

\item{...}{Option to pass on to \code{\link[tidyr:separate]{separate}}.}
}
\value{
Phyloseq with modified sample_data
}
\description{
Separate a Column in Phyloseq Sample Data
}
\details{
In some cases user may want to separate a sample column in
         \code{phyloseq} object. To ease this step, we make use of the
         \code{\link[tidyr:separate]{separate}}. The separation can be done
         with regular expression or numeric locations.
}
\examples{
library(biomeUtils)
library(dplyr)
data(FuentesIliGutData)
FuentesIliGutData <- mutateSampleData(FuentesIliGutData,
                                      column.to.split = paste0(sex,"_",age))
FuentesIliGutData <- separateSampleColumn(FuentesIliGutData,
                                          column.to.split, c("A", "B"), sep="_")

}
\references{
Shetty SA (2020). Utilities for microbiome analytics.
\url{https://github.com/microsud/biomeUtils}
}
\author{
Sudarshan A. Shetty
}
