% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeTaxonomicAssignments.R
\name{summarizeTaxonomicAssignments}
\alias{summarizeTaxonomicAssignments}
\title{Summarize Taxonomic Assignments}
\usage{
summarizeTaxonomicAssignments(x)
}
\arguments{
\item{x}{A phyloseq object}
}
\value{
tibble with overview of taxonomic assignments.
}
\description{
Summarize Taxonomic Assignments
}
\details{
Percentage of taxa that have been assigned to the different
         taxonomic levels. Non assigned levels must be NA. Check
         biomeUtils::cleanTaxonomy.
}
\examples{
library(biomeUtils)
data("SprockettTHData")
# run cleanTaxonomy if special characters are
# present in tax_table
summarizeTaxonomicAssignments(SprockettTHData)
}
\author{
Sudarshan A. Shetty
}
