% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformPhyloseq.R
\name{transformation}
\alias{transformation}
\alias{getProportions}
\alias{getCLR}
\title{Common Transformations}
\usage{
getProportions(x)

getCLR(x)
}
\arguments{
\item{x}{\code{phyloseq} object}
}
\value{
Transformed phyloseq
}
\description{
Common Transformations
}
\details{
Common transformation in microbiome like proportions, clr.

\code{getProportions} convert counts to relative abundance.

\code{getCLR} convert counts to centred log transform 'clr'
}
\examples{
library(biomeUtils)
data("FuentesIliGutData")
ps.rel <- getProportions(FuentesIliGutData)
ps.clr <- getCLR(FuentesIliGutData)

}
\references{
Shetty SA (2020). Utilities for microbiome analytics.
\url{https://github.com/RIVM-IIV-Microbiome/biomeUtils}
}
\author{
Sudarshan A. Shetty
}
