% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendAbundanceAssignment.R
\name{trendAbundanceAssignment}
\alias{trendAbundanceAssignment}
\title{Check Trend in Abundance and Taxonomic Assignment}
\usage{
trendAbundanceAssignment(x, quantiles = seq(0, 95, by = 10), plot = TRUE)
}
\arguments{
\item{x}{A phyloseq object}

\item{quantiles}{Abundances values to sort. Can be changed to specify
how many values in a distribution are above or below
a certain limit.}

\item{plot}{Logical. Default is TRUE.}
}
\value{
Either a list with data.frame and plot or just
        ggplot object.
}
\description{
Check Trend in Abundance and Taxonomic Assignment
}
\details{
Check if the more abundance-prevalent taxa
         have better taxonomic assignments. This is
         a pre-check, not corrected for differences
         in sequencing depth.
}
\examples{
library(biomeUtils)
data("SprockettTHData")

p1 <- trendAbundanceAssignment(SprockettTHData,
                              quantiles = seq(0, 95, by = 5),
                              plot=TRUE)
p1 + ggplot2::scale_colour_brewer("", palette = "Spectral") +
    ggplot2::theme_minimal() +
    ggplot2::theme(axis.text.x = element_text(angle=90, vjust=0.5))

}
\references{
\itemize{
\item{}{Thorsten Brach, 2018. MicrobiomeX2 Pipeline.
       \emph{GitHub}.
       \url{https://github.com/TBrach/MicrobiomeX2}
       \url{https://github.com/TBrach/Dada_Pipel/blob/master/Generalized_Phyloseq_Analysis_New.Rmd}
}
}
}
\author{
Sudarshan A. Shetty
}
