% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueness.R
\name{uniqueness}
\alias{uniqueness}
\title{Microbiota Uniqueness}
\usage{
uniqueness(x, dist_mat, reference_samples = NULL, ...)
}
\arguments{
\item{x}{A phyloseq object}

\item{dist_mat}{An object of class \code{dist}}

\item{reference_samples}{Vector of samples to use as reference.}

\item{...}{Option to pass biomeUtils::getSampleTibble like column name and columns
to be included in the output.}
}
\value{
A data frame with uniqueness and sample information
}
\description{
Microbiota Uniqueness
}
\details{
Extracts the minimum value from a \code{dissimilarity} matrix for each
         individual. This is the dissimilarity of an individual from their
         nearest neighbor. Here, the option of using a one or more reference
         samples is provided.
}
\examples{
library(biomeUtils)
data("FuentesIliGutData")
# Keep only two groups. controls and L1.
ps <- filterSampleData(FuentesIliGutData, ILI != "L2")
ps <- getProportions(ps)
dist.mat <- phyloseq::distance(ps, "bray")
# Define controls as reference samples
ref_samples <- rownames(meta(subset_samples(ps, ILI == "C")))
muniq <- uniqueness(ps,
                    dist_mat=dist.mat,
                    reference_samples = ref_samples)
head(muniq)
}
\references{
\itemize{
\item{}{Wilmanski T et al. (2021). Gut microbiome pattern reflects
healthy ageing and predicts survival in humans.
\emph{Nature metabolism}, 3(2), pp.274-286.}
}
}
\author{
Sudarshan A. Shetty
}
