% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniteGenusSpeciesNames.R
\name{uniteGenusSpeciesNames}
\alias{uniteGenusSpeciesNames}
\title{Unite Genus Species Names}
\usage{
uniteGenusSpeciesNames(x)
}
\arguments{
\item{x}{A phyloseq object}
}
\value{
A \code{phyloseq} with united genus and species
        in taxa_table .
}
\description{
Unite Genus Species Names
}
\details{
Unite genus and species names in tax_table \code{phyloseq}
         objects. E.g. \emph{Fusicatenibacter} and \emph{saccharivorans} are united
         to \emph{Fusicatenibacter.saccharivorans}.
}
\examples{
library(biomeUtils)
data("FuentesIliGutData")
ps <- uniteGenusSpeciesNames(FuentesIliGutData)

head(tax_table(ps)[, 7])
}
\references{
Shetty SA (2021). Utilities for microbiome analytics.
\url{https://github.com/microsud/biomeUtils}
}
\author{
Sudarshan A. Shetty
}
