% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniteSampleColumn.R
\name{uniteSampleColumn}
\alias{uniteSampleColumn}
\title{unite a Column in Phyloseq Sample Data}
\usage{
uniteSampleColumn(x, ...)
}
\arguments{
\item{x}{\code{phyloseq} object}

\item{...}{Option to pass on to \code{\link[tidyr:unite]{unite}}.}
}
\value{
Phyloseq with modified sample_data
}
\description{
unite a Column in Phyloseq Sample Data
}
\details{
In some cases user may want to unite a sample column in
         \code{phyloseq} object. To ease this step, we make use of the
         \code{\link[tidyr:unite]{unite}}. The separation can be done
         with regular expression or numeric locations.
}
\examples{
library(biomeUtils)
library(dplyr)
data(FuentesIliGutData)
FuentesIliGutData <- uniteSampleColumn(FuentesIliGutData,
                                       "united.column", sex:age,
                                       remove = FALSE)

}
\references{
Shetty SA (2020). Utilities for microbiome analytics.
\url{https://github.com/microsud/biomeUtils}
}
\author{
Sudarshan A. Shetty
}
